Added reverse geocoding for location value, and tweaks
