+++ client/public/index.html
@@ -5,6 +5,10 @@
         <meta name="theme-color" content="#000000" />
         <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
         <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+        <script
+            type="text/javascript"
+            src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAfKsvjQLoBQpRrWCEJraXSvMFnUunnOeI&libraries=places"
+        ></script>
     </head>
     <body>
         <div id="root"></div>
