+++ client/src/components/Services/ServicesFilter/ServicesFilter.tsx
@@ -24,6 +24,7 @@ const ServicesFilter = ({ handleFilterOption }: Props) => {
                 options={filterOptions}
                 styles={FilterStyles()}
                 onChange={handleFilterOption}
+                isSearchable={false}
             />
         </StyledServicesFilter>
     );
